fix: update NFT collection pages to use NFTCollectionView component // able to make changes that will work seamlessly across all NFTcollection pages
- Refactored DogePound, King, Rollbots, Sports Rollbots, Wilder World, and ZK Race collection pages to utilize the new NFTCollectionView component for improved consistency and maintainability.
- Removed unnecessary state management and UI logic from these pages, simplifying the codebase.
- Added imageUrl property to WilderWorldNFT interface for better image handling.
